requirements.txt

fastapi==0.115.6
uvicorn[standard]==0.34.0
azure-storage-blob==12.24.0
